requirements.txt
coverage==4.5.1
Django==2.1.2
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-extensions==2.1.3
django-recaptcha==1.4.0
django-timeline-logger==1.1.2
django-tinymce==2.7.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
flake8==3.5.0
icalendar==4.0.3
isort==4.3.4
jdcal==1.4
mccabe==0.6.1
odfpy==1.3.6
openpyxl==2.5.9
psycopg2==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
